q1_banking_assistant/requirements.txt
# Core LangChain and AI Framework
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-google-genai>=0.0.6

# Document Processing
unstructured>=0.12.0
pypdf>=4.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
pandas>=2.1.0
numpy>=1.24.0

# Vector Database
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Google AI (Gemini)
google-generativeai>=0.3.0

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
tiktoken>=0.5.0
tqdm>=4.66.0

# Development and Testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0 